add AbstractSQLBuilder
